% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc.R
\name{acc}
\alias{acc}
\alias{acc.clusterAssignment}
\alias{acc.seqnextCases}
\title{A Function to Compute the Accuracy of the Prediction}
\usage{
acc(obj, true.cluster)

\method{acc}{clusterAssignment}(obj, true.cluster)

\method{acc}{seqnextCases}(obj, true.cluster)
}
\arguments{
\item{obj}{an object of class "\code{clusterAssignment}".}

\item{true.cluster}{the true cluster of the new cases.}
}
\value{
A numeric or data frame depicting the accuracy of the prediction.
If the function is returned as data frame, it contains the following colums:
\itemize{
   \item{\code{cluster}:}{ a vector of the clusters' names in the test set.}
   \item{\code{accuracy}:}{ the fraction of correct assignment given cluster.}
}
}
\description{
A function to compute accuracy of the prediction.
}
\details{
If the \code{obj} is of class "\code{clusterAssignment}", the accuracy is
the fraction of true assignments within the predicted plausible clusters. For example,
if \code{nbest = 3} and the true cluster of a case is one of those 3 predicted clusters, then
the method assigns the case correctly. \cr \cr
If the \code{obj} is of class "\code{seqnextCases}", the accuracy is computed as the
fraction of cases which correctly are assigned to their true clusters. The accuracy of a cluster is
returned as \code{NaN} if there is no cases in such a cluster in the training set, but there is at
least one cases of such a cluster in the test set.
}
