% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case2sequence.R
\name{case2sequence}
\alias{case2sequence}
\title{Choose Which Cases to be Sequenced Next Given Clusters}
\usage{
case2sequence(obj, case.id = NULL, threshold = NULL, nbest = 3)
}
\arguments{
\item{obj}{an object of class "\code{clusterAssignment}".}

\item{case.id}{a (optional) vector of case id in which we should choose to be sequenced next.}

\item{threshold}{optional value to determine the cut-off probability, see "Details". Default is 0.}

\item{nbest}{how many predicted clusters to choose. Default is 3.}
}
\value{
By default, it shows data frame showing which cases to be sequenced next for a subset of clusters.
There are some values that can be extracted such as:
\itemize{
   \item{\code{$id}:}{ a vector of case id.}
   \item{\code{$best_cases}:}{ a data frame which determines which cases to be sequenced next for every cluster.}
}
}
\description{
A function to predict the most likely new cases to be sequenced next given a cluster. It ranks
the cases based on their priorities in each cluster.
}
\details{

}
