% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterPLR.R
\name{clusterPLR}
\alias{clusterPLR}
\title{Cluster Assignment Method}
\usage{
clusterPLR(formula, data, newdata, threshold = NULL, nbest = 3)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}}. The details of model specification are given under ‘Details’.}

\item{data}{a data frame containing the variables in the model. If geographic location is included into the model, name the column to "lat" (or "latitude") as reference to the latitude and "long" (or "longitude") as reference to the longitude.}

\item{newdata}{a data frame containing the variables of some cases in which we want to predict their clusters.}

\item{threshold}{optional value to determine the cut-off probability, see "Details".}

\item{nbest}{how many predicted clusters to choose. Default is 3.}
}
\value{
By default, it returns a data frame showing the \code{nbest} clusters for the first six cases.
Some value can be extracted, such as
\itemize{
   \item{\code{$best_cluster}:}{ a data frame showing the best predicted clusters given \code{threshold} and/or \code{nbest} for every case.
   See \code{?getbest} for more details on its data frame.}
   \item{\code{$cluster_score}:}{ a data frame showing the predicted probabilities or scores of a case to be assigned into each cluster.
   See \code{?getclusterScore} for more details on its data frame.}
   \item{\code{$model_fit}}:{ summary of the model fit using multinomial logistic regression.}
}
}
\description{
A function to predict a case's cluster using pairwise logistic regression.
}
\details{
\code{formula} argument is in the format of \code{response ~ predictors}.
The "response" is a character/factor vector of the clusters' names or indexes. \cr \cr
If \code{threshold} is not provided, by default the method will choose the \code{nbest}
predicted clusters.
}
\seealso{
Other clusterAssignment: 
\code{\link{clusterMLR}()},
\code{\link{clusterRandom}()}
}
\concept{clusterAssignment}
