% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterRandom.R
\name{clusterRandom}
\alias{clusterRandom}
\title{Cluster Assignment Method}
\usage{
clusterRandom(x, N = 1, threshold = NULL, nbest = 3)
}
\arguments{
\item{x}{a (character or factor) vector of the true clusters in the training data.}

\item{N}{number of cases in which we want to predict their clusters. Default is 1.}

\item{threshold}{optional value to determine the cut-off probability, see "Details".}

\item{nbest}{how many predicted clusters to choose. Default is 3.}
}
\value{
By default, it returns a data frame showing the \code{nbest} clusters for the first six cases.
Some value can be extracted, such as
\itemize{
   \item{\code{$best_cluster}:}{ a data frame showing the best predicted clusters given \code{threshold} and/or \code{nbest} for every case.
   See \code{?getbest} for more details on its data frame.}
   \item{\code{$cluster_score}:}{ a data frame showing the predicted probabilities or scores of a case to be assigned into each cluster.
   See \code{?getclusterScore} for more details on its data frame.}
}
}
\description{
A function to assign cases in clusters randomly, proportional to the cluster's size.
}
\details{
If \code{threshold} is not provided, by default the method will choose the \code{nbest}
predicted clusters randomly.
}
\seealso{
Other clusterAssignment: 
\code{\link{clusterMLR}()},
\code{\link{clusterPLR}()}
}
\concept{clusterAssignment}
