% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterSize.R
\name{clusterSize}
\alias{clusterSize}
\alias{clusterSize.clusterAssignment}
\title{Estimate the Cluster Size Increment}
\usage{
clusterSize(obj, rho = 0)

\method{clusterSize}{clusterAssignment}(obj, rho = 0)
}
\arguments{
\item{obj}{an object of class "\code{clusterAssignment}".}

\item{rho}{a probability that a case does not belong to any of the given clusters.}
}
\value{
A data frame of the estimated increment of clusters.
}
\description{
A function to estimate the cluster size increment after assigning unsequenced cases.
}
