% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbest.R
\name{getbest}
\alias{getbest}
\alias{getbest.clusterAssignment}
\alias{getbest.seqnextCases}
\title{Extract the Best Cluster/Cases}
\usage{
getbest(obj)

\method{getbest}{clusterAssignment}(obj)

\method{getbest}{seqnextCases}(obj)
}
\arguments{
\item{obj}{an object of class "\code{clusterAssignment}" or "\code{seqnextCases}".}
}
\value{
A data frame of the best cluster(s) assignment or the cases which should be sequenced next,
depending on the \code{obj}'s class.
\itemize{
   \item{\code{rank_n}:}{ a column showing the n-th best cluster of a case, ranked by its cluster scores.}
   \item{\code{cluster}:}{ cluster's name or index.}
   \item{\code{priority_n}:}{ a column showing the n-th best case most likely belongs to a cluster.}
}
}
\description{
A function to extract a table of the best probable clusters or
the predicted cases, ranked by their priorities.
}
