% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getclusterScore.R
\name{getclusterScore}
\alias{getclusterScore}
\alias{getclusterScore.clusterAssignment}
\title{Get Cluster Score}
\usage{
getclusterScore(obj)

\method{getclusterScore}{clusterAssignment}(obj)
}
\arguments{
\item{obj}{an object of class "\code{clusterAssignment}".}
}
\value{
A data frame of the new cases' cluster scores.
}
\description{
A function to extract the cluster scores of unsequenced cases in each cluster.
}
