% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_constructor.R
\name{new_optimumThreshold}
\alias{new_optimumThreshold}
\title{Constructor to define optimumThreshold class}
\usage{
new_optimumThreshold(x)
}
\description{
Constructor to define optimumThreshold class
}
