% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plr.R
\name{plr}
\alias{plr}
\title{Fit Pairwise Logistic Regression}
\usage{
plr(formula, data)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}}. The details of model specification are given under "Details".}

\item{data}{a data frame containing the variables in the model. If geographic location is included into the model, name the
column to "lat" (or "latitude") as reference to the latitude and "long" (or "longitude") as reference to the longitude.}
}
\value{
Returns an object of class "\code{plr}", which inherits from "glm" class and "lm" class.
}
\description{
A function to fit a pairwise logistic regression.
}
\details{
\code{formula} argument is in the format of \code{response ~ predictors}.
The "response" is a character/factor vector of the clusters' names or indexes.
}
\examples{

}
