% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.plr.R
\name{predict.plr}
\alias{predict.plr}
\title{Model Prediction}
\usage{
\method{predict}{plr}(obj, newdata, case.id = NULL)
}
\arguments{
\item{obj}{a fitted object of a class from "\code{plr}".}

\item{newdata}{a data frame containing the variables of some cases in which we want to predict.}

\item{case.id}{optional vector of case id or index in the \code{newdata}.}
}
\value{
a data frame containing:
\itemize{
   \item{\code{y}: }{the predicted probability if \code{case-to.case} is in the same cluster.}
   \item{\code{se.y:} }{the standard error of \code{y}.}
}
}
\description{
Obtains predictions and estimates standard errors from a fitted pairwise logistic regression model object.
}
